Virtual Private Networks VPN's
Observing the Effect of a VPN on IP Addressing
-------------------------------------------------------------------------
In this project, I observed the impact of a virtual private network (VPN) on IP address, location, and web browsing. The video will guide you visually through the steps involved and then further on there are some notes and screenshots provided for highlights of some or all of the material.